add run wasm function feature in playground (#389)

+++ ts2wasm/scripts/link_hooks.ts
@@ -15,14 +15,16 @@ hooks.forEach(function (hook) {
     const hookInSourceControl = path.resolve(__dirname, 'hooks', hook);
 
     if (fs.existsSync(hookInSourceControl)) {
+        const gitHookDirectory = path.resolve(
+            __dirname, '..', '..', '.git', 'hooks');
         const hookInHiddenDirectory = path.resolve(
-            __dirname,
-            '..',
-            '..',
-            '.git',
-            'hooks',
+            gitHookDirectory,
             hook,
         );
+
+        if (!fs.existsSync(gitHookDirectory)) {
+            return;
+        }
 
         if (fs.existsSync(hookInHiddenDirectory)) {
             fs.unlinkSync(hookInHiddenDirectory);
